Time: 9 ms (10.25%), Space: 5.9 MB (75.35%) - LeetHub

+++ 0172-factorial-trailing-zeroes/0172-factorial-trailing-zeroes.cpp
@@ -3,7 +3,7 @@ public:
 int trailingZeroes(int n)
 {
     int res=0;
-    unsigned long long x=5;
+    int x=5;
     while((n/x)>0)
     {
         res+=(n/x);
